Renamed into model package - Model Design https://github.com/sowbharni-05/E-ComRest-Impl/issues/3#issue-910471175
