--- src/main/java/com/banfico/EcomApplication/Model/Category.java
@@ -1,25 +1,0 @@
-package com.banfico.EcomApplication.Model;
-
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.List;
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@Entity
-@Table(name="category")
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private int categoryId;
-    private String CategoryName;
-    @JsonManagedReference
-    @OneToMany(targetEntity = Product.class,mappedBy = "category")
-    private List<Product> products;
-}
