--- src/main/java/com/banfico/EcomApplication/Model/Product.java
@@ -1,28 +1,0 @@
-package com.banfico.EcomApplication.Model;
-
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-@Entity
-@Table(name = "product")
-public class Product {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private int productId;
-    private String productName;
-    private double price;
-    @JsonBackReference
-    @ManyToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "categoryId",referencedColumnName ="categoryId")
-    private Category category;
-
-}
