+++ src/main/java/com/banfico/EcomApplication/model/Category.java
@@ -1,0 +1,19 @@
+package com.banfico.EcomApplication.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+
+@Getter
+@Setter
+@MappedSuperclass
+public class Category {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private int categoryId;
+
+}
