+++ src/main/java/com/banfico/EcomApplication/model/Product.java
@@ -1,0 +1,18 @@
+package com.banfico.EcomApplication.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+
+@MappedSuperclass
+@Setter
+@Getter
+public class Product {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private int productId;
+}
